Add GHA to test AsciiDoc builds on each PR
Add a GitHub action that builds each .adoc file on each PR, failing if
there are any errors above WARN-level.

Shared automated testing infra makes the contribution process more
comfortable for developers and the code review process easier for
owners. It also guarantees consistent standards across the project.

Subsequent PRs should lower the threshold for failing to WARN and fix
issues, and then lower it again to INFO and fix issues.

This can eventually be extracted to a stand-alone GitHub action,
perhaps moved to a different repo, and shared with the rest of the
AsciiDoc ecosystem.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>